+++ models/userModel.js
@@ -2,20 +2,23 @@
 
 // access database
 const pool = require('../database/db');
+const { httpError } = require('../utils/errors');
 const promisePool = pool.promise();
 
 // get all users
-const getAllUsers = async () => {
+const getAllUsers = async (next) => {
   try {
     const [rows] = await promisePool.execute('SELECT * FROM insign_user');
     return rows;
   } catch (error) {
-    console.log(error.message);
+    console.error('model get all users', e.message);
+    const err = httpError('Sql error', 500);
+    next(err);
   }
 };
 
 // get user by Id
-const getUser = async (userId) => {
+const getUser = async (userId, next) => {
   try {
     const [rows] = await promisePool.execute(
       'SELECT * FROM insign_user where user_id = ?',
@@ -23,12 +26,14 @@ const getUser = async (userId) => {
     );
     return rows[0];
   } catch (error) {
-    console.log(error.message);
+    console.error('model get user by id', e.message);
+    const err = httpError('Sql error', 500);
+    next(err);
   }
 };
 
 // add new user
-const insertUser = async (user) => {
+const insertUser = async (user, next) => {
   try {
     const [rows] = await promisePool.execute(
       'INSERT INTO insign_user(username, email, password, profile_picture, bio, role_id) VALUES(?, ?, ?, ?, ?, ?)',
@@ -42,26 +47,34 @@ const insertUser = async (user) => {
       ]
     );
     return rows;
-  } catch (error) {
-    console.log(error.message);
+  } catch (e) {
+    console.error('model insert user', e.message);
+    const err = httpError('Sql error', 500);
+    next(err);  
   }
 };
 
 // delete user
-const deleteUser = async (userId) => {
+const deleteUser = async (userId, role_id, next) => {
+  let sql = 'DELETE FROM insign_user WHERE user_id = ? AND role_id = ?'
+  let params = [userId, role_id];
+  // admin can delete user
+  if(role_id === 0) {
+    sql = 'DELETE FROM insign_user WHERE user_id = ?';
+    params = [userId]
+  }
   try {
-    const [rows] = await promisePool.execute(
-      'DELETE FROM insign_user WHERE user_id = ?',
-      [userId]
-    );
+    const [rows] = await promisePool.execute(sql, params);
     return rows.affectedRows === 1;
-  } catch (error) {
-    console.log(error.message);
+  } catch (e) {
+    console.error('model delete user', e.message);
+    const err = httpError('Sql error', 500);
+    next(err);  
   }
 };
 
 // update user
-const updateUser = async (userId, user) => {
+const updateUser = async (userId, user, next) => {
   try {
     const [rows] = await promisePool.execute(
       'UPDATE insign_user SET username = ?, email = ?, password = ?, profile_picture = ?, bio= ?, role_id = ? WHERE user_id = ?',
@@ -78,11 +91,13 @@ const updateUser = async (userId, user) => {
     return rows.affectedRows === 1;
   } catch (e) {
     console.error('model update user', e.message);
+    const err = httpError('Sql error', 500);
+    next(err);  
   }
 };
 
 //get all Posts of a user with number of comments and likes
-const getAllPostsOfUser = async (userId) => {
+const getAllPostsOfUser = async (userId, next) => {
   try {
     const [rows] = await promisePool.execute(
       'SELECT post.post_id, post.title, post.image, (SELECT count(*) from likes WHERE likes.post_id = post.post_id) as num_likes, (SELECT count(*) from comment WHERE comment.post_id = post.post_id) as num_comments FROM post WHERE post.author = ?',
@@ -90,7 +105,21 @@ const getAllPostsOfUser = async (userId) => {
     );
     return rows;
   } catch (error) {
-    console.log(error.message);
+    const err = httpError('Sql error', 500);
+    next(err);
+  }
+};
+
+// user log in
+const getUserLogin = async (params) => {
+  try {
+    console.log(params);
+    const [rows] = await promisePool.execute(
+        'SELECT * FROM insign_user WHERE email = ?;',
+        params);
+    return rows;
+  } catch (e) {
+    console.log('error', e.message);
   }
 };
 
@@ -101,4 +130,5 @@ module.exports = {
   deleteUser,
   updateUser,
   getAllPostsOfUser,
+  getUserLogin,
 };
